requirements.txt
python-chess==1.999
numpy==1.26.4
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.67.1
matplotlib==3.9.2
pandas==2.2.1
tqdm==4.67.1
seaborn==0.13.2
pillow==10.4.0
cairosvg==2.7.1
berserk